Changement des couleurs lorsque l'image ne contient que 2 couleurs
